Setup the Initial Layout of the Frontend Project
This included adding the bootstrap to the project packages and adding
it to the imports, as well as defining a few React Elements and setting
some hard-coded CSS.

Next steps are to actually build in a backend endpoint that can spit
out data appropriate for this app. That's going to involve extracting
functionality from my `dnsanalyzergo` project into a library. I also
plan to mess around with GAE for deployment, but may go a different
route due to database fees.