Added some component tests and made them pass
I need a new commit & push to trigger travis, and I figured why not
mess around with the testing. It's pretty cool and I figured out that I
should make sure components should be able to handle the case where
they don't get passed any properties. I also had to export more
components to use them in the test suite.

I plan to come back here and split out the components and testing to be
far more organized than it currently is, but I think I want to move the
table state into a property and pass the state in from the top of the
app, and then build the serverside portion which actually allows the
data to pour into the app so I can watch it work for real.