Made `flow check` return no errors
I made a bunch of changes and left todo notes all over the place. This
is another ugly commit which had the sole goal of getting flow type
checks to pass So I had to eradicate the use of an objeet and replace
it with a type instead and some other stuff.